requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
gunicorn==22.0.0
python-dotenv==1.0.1
psycopg[binary]==3.2.13
email-validator==2.2.0